feat: refactor impl for all plugins
Refactor Impl for all plugins.

Nddtfjiang <zhicheng.jiang@merico.dev>
