+++ plugins/icla/plugin_main.go
@@ -18,94 +18,13 @@ limitations under the License.
 package main
 
 import (
-	"fmt"
-
-	"github.com/apache/incubator-devlake/errors"
-	"github.com/apache/incubator-devlake/plugins/core"
-	"github.com/apache/incubator-devlake/plugins/helper"
-	"github.com/apache/incubator-devlake/plugins/icla/models"
-	"github.com/apache/incubator-devlake/plugins/icla/models/migrationscripts"
-	"github.com/apache/incubator-devlake/plugins/icla/tasks"
+	"github.com/apache/incubator-devlake/plugins/icla/impl"
 	"github.com/apache/incubator-devlake/runner"
 	"github.com/spf13/cobra"
-	"github.com/spf13/viper"
-	"gorm.io/gorm"
 )
 
-// make sure interface is implemented
-var _ core.PluginMeta = (*Icla)(nil)
-var _ core.PluginInit = (*Icla)(nil)
-var _ core.PluginTask = (*Icla)(nil)
-var _ core.PluginApi = (*Icla)(nil)
-var _ core.PluginModel = (*Icla)(nil)
-var _ core.PluginMigration = (*Icla)(nil)
-var _ core.CloseablePluginTask = (*Icla)(nil)
-
 // PluginEntry is a variable exported for Framework to search and load
-var PluginEntry Icla //nolint
-
-type Icla struct{}
-
-func (plugin Icla) Description() string {
-	return "collect some Icla data"
-}
-
-func (plugin Icla) Init(config *viper.Viper, logger core.Logger, db *gorm.DB) errors.Error {
-	return nil
-}
-
-func (plugin Icla) GetTablesInfo() []core.Tabler {
-	return []core.Tabler{
-		&models.IclaCommitter{},
-	}
-}
-
-func (plugin Icla) SubTaskMetas() []core.SubTaskMeta {
-	return []core.SubTaskMeta{
-		tasks.CollectCommitterMeta,
-		tasks.ExtractCommitterMeta,
-	}
-}
-
-func (plugin Icla) PrepareTaskData(taskCtx core.TaskContext, options map[string]interface{}) (interface{}, errors.Error) {
-	var op tasks.IclaOptions
-	err := helper.Decode(options, &op, nil)
-	if err != nil {
-		return nil, err
-	}
-
-	apiClient, err := errors.Convert01(tasks.NewIclaApiClient(taskCtx))
-	if err != nil {
-		return nil, err
-	}
-
-	return &tasks.IclaTaskData{
-		Options:   &op,
-		ApiClient: apiClient,
-	}, nil
-}
-
-// PkgPath information lost when compiled as plugin(.so)
-func (plugin Icla) RootPkgPath() string {
-	return "github.com/apache/incubator-devlake/plugins/icla"
-}
-
-func (plugin Icla) MigrationScripts() []core.MigrationScript {
-	return migrationscripts.All()
-}
-
-func (plugin Icla) ApiResources() map[string]map[string]core.ApiResourceHandler {
-	return nil
-}
-
-func (plugin Icla) Close(taskCtx core.TaskContext) errors.Error {
-	data, ok := taskCtx.GetData().(*tasks.IclaTaskData)
-	if !ok {
-		return errors.Default.New(fmt.Sprintf("GetData failed when try to close %+v", taskCtx))
-	}
-	data.ApiClient.Release()
-	return nil
-}
+var PluginEntry impl.Icla //nolint
 
 // standalone mode for debugging
 func main() {
