+++ plugins/starrocks/starrocks.go
@@ -18,51 +18,12 @@ limitations under the License.
 package main
 
 import (
-	"github.com/apache/incubator-devlake/errors"
-	"github.com/apache/incubator-devlake/plugins/core"
-	"github.com/apache/incubator-devlake/plugins/helper"
+	"github.com/apache/incubator-devlake/plugins/starrocks/impl"
 	"github.com/apache/incubator-devlake/runner"
 	"github.com/spf13/cobra"
 )
 
-type StarRocks string
-
-// make sure interface is implemented
-var _ core.PluginMeta = (*StarRocks)(nil)
-var _ core.PluginTask = (*StarRocks)(nil)
-var _ core.PluginModel = (*StarRocks)(nil)
-
-func (s StarRocks) SubTaskMetas() []core.SubTaskMeta {
-	return []core.SubTaskMeta{
-		LoadDataTaskMeta,
-	}
-}
-
-func (s StarRocks) PrepareTaskData(taskCtx core.TaskContext, options map[string]interface{}) (interface{}, errors.Error) {
-	var op StarRocksConfig
-	err := helper.Decode(options, &op, nil)
-	if err != nil {
-		return nil, err
-	}
-	if op.BeHost == "" {
-		op.BeHost = op.Host
-	}
-	return &op, nil
-}
-
-func (s StarRocks) GetTablesInfo() []core.Tabler {
-	return []core.Tabler{}
-}
-
-func (s StarRocks) Description() string {
-	return "Sync data from database to StarRocks"
-}
-
-func (s StarRocks) RootPkgPath() string {
-	return "github.com/merico-dev/lake/plugins/starrocks"
-}
-
-var PluginEntry StarRocks
+var PluginEntry impl.StarRocks
 
 func main() {
 	cmd := &cobra.Command{Use: "StarRocks"}
