+++ plugins/starrocks/tasks/task_data.go
@@ -15,7 +15,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package main
+package tasks
 
 type StarRocksConfig struct {
 	SourceType  string `mapstructure:"source_type"`
