+++ plugins/starrocks/tasks/tasks.go
@@ -15,7 +15,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package main
+package tasks
 
 import (
 	"bytes"
@@ -32,6 +32,7 @@ import (
 	"github.com/apache/incubator-devlake/impl/dalgorm"
 	"github.com/apache/incubator-devlake/plugins/core"
 	"github.com/apache/incubator-devlake/plugins/core/dal"
+	"github.com/apache/incubator-devlake/plugins/starrocks/utils"
 	"github.com/lib/pq"
 	"gorm.io/driver/mysql"
 	"gorm.io/driver/postgres"
@@ -76,7 +77,7 @@ func LoadData(c core.SubTaskContext) errors.Error {
 	}
 	var starrocksTables []string
 	if config.DomainLayer != "" {
-		starrocksTables = getTablesByDomainLayer(config.DomainLayer)
+		starrocksTables = utils.GetTablesByDomainLayer(config.DomainLayer)
 		if starrocksTables == nil {
 			return errors.NotFound.New(fmt.Sprintf("no table found by domain layer: %s", config.DomainLayer))
 		}
@@ -183,7 +184,7 @@ func createTmpTable(starrocks *sql.DB, db dal.Dal, starrocksTmpTable string, tab
 		if !ok {
 			return columnMap, "", errors.Default.New(fmt.Sprintf("Get [%s] ColumeType Failed", name))
 		}
-		dataType := getStarRocksDataType(columnDatatype)
+		dataType := utils.GetStarRocksDataType(columnDatatype)
 		columnMap[name] = dataType
 		column := fmt.Sprintf("`%s` %s", name, dataType)
 		columns = append(columns, column)
